hostapd: revert upstream commit 01677c47f to fix crashes
Commit 01677c47f ("AP: Support disconnect with MLD") appears to
be causing a SIGSEGV in some cases for EAP APs when a client
disconnects from the AP. As the core dump doesn't appear to be
useful and the solution isn't immediately obvious, we revert
the upstream commit until a better solution could be found.

Credit to e9hack[1] for identifying the commit causing this
issue. The full discussion can be found in this PR[2].

[1]: https://github.com/e9hack
[2]: https://github.com/openwrt/openwrt/pull/17423

Signed-off-by: Rany Hany <rany_hany@riseup.net>
(cherry picked from commit fc8b0677c8362e6818d8bec60af1c999fafe4b0b)